extract ecto related function from view into lib
